refactor (Event): add class Screen containing the event listeners
